requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chromedriver==2.24.1
chromedriver-binary==106.0.5249.61.0
click==8.1.3
exceptiongroup==1.0.0rc9
firebase-admin==6.0.0
get-chrome-driver==1.3.9
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
h11==0.14.0
httplib2==0.20.4
idna==3.4
msgpack==1.0.4
outcome==1.2.0
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
requests==2.28.1
rsa==4.9
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
typer==0.6.1
uritemplate==4.1.1
urllib3==1.26.12
wsproto==1.2.0
